requirements.txt
flask
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
flask-admin
enum
flask-restful
requests
nose2